Renamed/moved some things in anticipation of future work
